requirements.txt
streamlit
openai
pypdf2
pandas
python-dotenv
